refactor parquet primitive types to be generic
